requirements.txt
streamlit
python-dotenv
transformers
sentence-transformers
faiss-cpu
langchain
